Feature: Change bucket texture depending on axolotl variant
This commit contains some code which could be better, but I have
been struggling to find a way to do it better than registering fake
items. If anyone has a better idea, please feel free to fix it.

My original plan was to just do ``itemModels.getModel(new ModelIdentifier("axolotlviewer:axolotl_bucket_blue"));``
but that doesn't seem to want to work, the best way I could get it
working was with "fake items". Which don't show in the creative menu as
they do not have a category. The only thing they for is to be able to
get the BakedModel from it.