Improvement: Stop using fake items to load texture models
Prior to this commit, the mod would register a fake item to force
Minecraft to load a model. This approach was quite messy, and not ideal.

This commit uses a ModelLoader mixin to tell Minecraft about our item
model, without having to register an Item.

The AxolotlViewer class has also been changed to a ClientModInitializer
for clarity.